+++ packages/medusa-plugin-contentful/src/services/contentful.js
@@ -449,9 +449,14 @@ class ContentfulService extends BaseService {
       "options",
     ]
 
-    const found = variant.fields.find((f) => updateFields.includes(f))
-    if (!found) {
-      return
+    // Update came directly from product variant service so only act on a couple
+    // of fields. When the update comes from the product we want to ensure 
+    // references are set up correctly so we run through everything.
+    if (variant.fields) {
+      const found = variant.fields.find((f) => updateFields.includes(f))
+      if (!found) {
+        return
+      }
     }
 
     try {